Handle tar files and show the multiqc report

+++ helpers/file_renaming.py
@@ -28,25 +28,30 @@ def extract_uploaded_gzip(uploaded_file_path, extract_directory):
     # Get the file name from the full path
     uploaded_file_name = os.path.basename(uploaded_file_path)
 
-    # Check if the uploaded file name ends with '.gz'
-    if not uploaded_file_name.endswith('.gz'):
-        raise ValueError("The uploaded file is not a gzip file.")
+    # Check if the uploaded file name ends with '.tar'
+    if uploaded_file_name.endswith('.tar'):
+        extract_tar_without_structure(uploaded_file_path, extract_directory)
+        #os.remove(uploaded_file_name)
+    else:
+        # Check if the uploaded file name ends with '.gz'
+        if not uploaded_file_name.endswith('.gz'):
+            raise ValueError("The uploaded file is not a gzip file.")
 
-    # Get the file name without the '.gz' extension
-    file_name = os.path.splitext(uploaded_file_name)[0]
+        # Get the file name without the '.gz' extension
+        file_name = os.path.splitext(uploaded_file_name)[0]
 
-    # Path to extract the file to
-    extract_path = os.path.join(extract_directory, file_name)
+        # Path to extract the file to
+        extract_path = os.path.join(extract_directory, file_name)
 
-    # Extract the gzip file
-    with gzip.open(uploaded_file_path, 'rb') as f_in:
-        with open(extract_path, 'wb') as f_out:
-            f_out.write(f_in.read())
+        # Extract the gzip file
+        with gzip.open(uploaded_file_path, 'rb') as f_in:
+            with open(extract_path, 'wb') as f_out:
+                f_out.write(f_in.read())
 
-    # check if the file is a tar, in which case we have more to do
-    if file_name.endswith('.tar'):
-        extract_tar_without_structure(extract_path, extract_directory)
-        os.remove(extract_path)
+        # check if the file is a tar, in which case we have more to do
+        if file_name.endswith('.tar'):
+            extract_tar_without_structure(extract_path, extract_directory)
+            os.remove(extract_path)
     
     return True
